+++ include/arion/common/threading_manager.hpp
@@ -2,10 +2,10 @@
 #define ARION_THREADING_MANAGER_HPP
 
 #include <arion/common/global_defs.hpp>
+#include <arion/unicorn/unicorn.h>
 #include <map>
 #include <memory>
 #include <stack>
-#include <arion/unicorn/unicorn.h>
 #include <vector>
 
 class Arion;
@@ -33,9 +33,8 @@ struct ARION_EXPORT ARION_THREAD
     arion::ADDR child_tid_addr;
     arion::ADDR parent_tid_addr;
     std::unique_ptr<std::map<arion::REG, arion::RVAL>> regs_state = nullptr;
-    int64_t lock = 0;
     arion::ADDR wait_status_addr = 0;
-    bool paused = false;
+    bool stopped = false;
     arion::ADDR robust_list_head = 0;
     arion::ADDR rseq_addr = 0;
     uint32_t rseq_len = 0;
@@ -50,7 +49,7 @@ struct ARION_EXPORT ARION_THREAD
           child_tid_addr(arion_t->child_tid_addr), parent_tid_addr(arion_t->parent_tid_addr),
           regs_state(arion_t->regs_state ? std::make_unique<std::map<arion::REG, arion::RVAL>>(*arion_t->regs_state)
                                          : nullptr),
-          lock(arion_t->lock), wait_status_addr(arion_t->wait_status_addr), paused(arion_t->paused) {};
+          wait_status_addr(arion_t->wait_status_addr), stopped(arion_t->stopped) {};
 };
 std::vector<arion::BYTE> serialize_arion_thread(ARION_THREAD *arion_t);
 ARION_THREAD *deserialize_arion_thread(std::vector<arion::BYTE> srz_thread);
@@ -66,26 +65,16 @@ struct ARION_EXPORT ARION_TGROUP_ENTRY
 class ARION_EXPORT ThreadingManager
 {
   private:
-    static std::map<int, std::string> sync_signals;
     std::weak_ptr<Arion> arion;
     pid_t curr_id = 1;
     pid_t running_tid = 1;
     std::stack<pid_t> free_thread_ids;
-    std::map<int, std::shared_ptr<struct ksigaction>> sighandlers;
-    static void intr_hook(std::shared_ptr<Arion> arion, uint32_t intno, void *user_data);
-    static bool invalid_memory_hook(std::shared_ptr<Arion> arion, uc_mem_type access, uint64_t addr, int size,
-                                    int64_t val, void *user_data);
-    static bool invalid_insn_hook(std::shared_ptr<Arion> arion, void *user_data);
     pid_t gen_next_id();
-    void handle_sync_signals();
-    void handle_async_signals();
-    void handle_wait_signals();
 
   public:
     static std::map<pid_t, std::vector<std::unique_ptr<ARION_TGROUP_ENTRY>>> thread_groups;
     std::map<pid_t, std::unique_ptr<ARION_THREAD>> threads_map;
     std::map<arion::ADDR, std::vector<std::unique_ptr<ARION_FUTEX>> *> futex_list;
-    std::map<pid_t, pid_t> sigwait_list;
     static std::unique_ptr<ThreadingManager> initialize(std::weak_ptr<Arion> arion);
     ThreadingManager(std::weak_ptr<Arion> arion);
     ~ThreadingManager();
@@ -98,12 +87,8 @@ class ARION_EXPORT ThreadingManager
     void switch_to_next_thread();
     void futex_wait(pid_t tid, arion::ADDR futex_addr, uint32_t futex_bitmask);
     void futex_wait_curr(arion::ADDR futex_addr, uint32_t futex_bitmask);
-    bool signal_wait(pid_t tid, pid_t target_pid);
-    bool signal_wait_curr(pid_t target_pid);
-    bool has_sighandler(int signo);
-    std::shared_ptr<struct ksigaction> get_sighandler(int signo);
-    void set_sighandler(int signo, std::shared_ptr<struct ksigaction> sighandler);
-    void handle_signals();
+    bool signal_wait(pid_t target_tid, pid_t source_pid, arion::ADDR wait_status_addr);
+    bool signal_wait_curr(pid_t source_pid, arion::ADDR wait_status_addr);
     size_t futex_wake(arion::ADDR futex_addr, uint32_t futex_bitmask);
     bool ARION_EXPORT is_curr_locked();
     size_t ARION_EXPORT get_threads_count();